Update patrolV2.mzn
This version included "globals.mzn" which is actually only needed in patrolV3.mzn. Removed for clarity.

Also

reformatted the constraints so that they look exactly as in patrolV1, no changes.
Shortened the long output command.

Interesting:

If the constraint about "not three night shifts" comes before the constraint "evening then not night" (in this change request), then a solution is found in 13s (on my machine).
If the constraint about "evening then not night" comes before the constraint "not three night shifts" (originally), then a solution is found in 33s (on my machine).